requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.1
google-generativeai>=0.3.2
openai>=1.12.0
python-dotenv>=1.0.0
plotly>=5.18.0
scipy>=1.12.0
SQLAlchemy
pytest
requests
